+++ src/core/index.ts
@@ -1,4 +1,4 @@
 export { BQueryCollection } from './collection';
 export { BQueryElement } from './element';
 export { $, $$ } from './selector';
-export { utils } from './utils';
+export * from './utils';